requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
python-telegram-bot==20.7
firebase-admin==6.2.0